Context overhaul (#61)
Overhaul API for React.createContext()

Several API changes were required to update to the new API. I also updated the tooling to initially be similar to react-access, however we ended up moving to rollup which seems to be better suited for libraries. Largely, copied some of this over from another library (react-request) to get things working. This covers the main areas of #53 and includes a 4.0.0-beta.1 release on npm.